[Feat] dockerize reciept service #49
receipt AI flask test를 위해POST method /receipt API 작성
form-data로 imageId 전송
flask에서 parsing 필요
