+++ src/service/nestjs/package.json
@@ -21,6 +21,7 @@
 	},
 	"dependencies": {
 		"@miaooo/class-transformer-trim": "^1.0.0",
+		"@nestjs/axios": "^3.0.0",
 		"@nestjs/common": "^10.0.0",
 		"@nestjs/config": "^3.0.1",
 		"@nestjs/core": "^10.0.0",
@@ -33,6 +34,7 @@
 		"@types/cookie-parser": "^1.4.4",
 		"@types/passport": "^1.0.12",
 		"@types/passport-jwt": "^3.0.9",
+		"axios": "^1.5.0",
 		"bcrypt": "^5.1.1",
 		"class-transformer": "^0.5.1",
 		"class-validator": "^0.14.0",
@@ -83,5 +85,6 @@
 		],
 		"coverageDirectory": "../coverage",
 		"testEnvironment": "node"
-	}
+	},
+	"proxy": "http://receipt:5000"
 }
