+++ src/service/nestjs/src/api/receipt/receipt.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { receiptController } from './receipt.controller';
+import { HttpModule } from '@nestjs/axios';
+
+@Module({
+	imports: [HttpModule],
+	controllers: [receiptController],
+})
+export class receiptModule {}
