+++ src/service/nginx/tool/setup.sh
@@ -8,8 +8,6 @@ sed -i "s|my_domain|$DOMAIN_NAME|g" /etc/nginx/sites-available/default
 sed -i "s|my_certs_key|$CERTS_KEY|g" /etc/nginx/sites-available/default
 sed -i "s|my_certs|$CERTS|g" /etc/nginx/sites-available/default
 
-cat /etc/nginx/sites-available/default
-
 #sed -i '40s/.*/\tlog_format debug '\''$remote_addr - $remote_user [$time_local] '\'' '\''"$request" $status $body_bytes_sent '\'' '\''"$http_referer" "$http_user_agent"'\'';/' /etc/nginx/nginx.conf
 #sed -i '41s/.*/\taccess_log \/var\/log\/nginx\/access.log debug;/' /etc/nginx/nginx.conf
 #sed -i '42s/.*/\terror_log \/var\/log\/nginx\/error.log debug;/' /etc/nginx/nginx.conf
