+++ src/service/receipt/.env.sample
@@ -1,0 +1,3 @@
+PORT=
+CLOVA_API_URL=
+CLOVA_API_SECRET_KEY=
