+++ src/service/receipt/src/templates/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<title>이미지 업로드</title>
+	</head>
+	<body>
+		<h1>이미지 업로드</h1>
+		<form action="/receipt" method="POST" enctype="multipart/form-data">
+			<input type="file" name="image" />
+			<input type="submit" value="업로드" />
+		</form>
+	</body>
+</html>
