--- src/service/reciept/.env.sample
@@ -1,1 +1,0 @@
-PORT=
