--- src/service/reciept/templates/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>이미지 업로드</title>
-  </head>
-  <body>
-    <h1>이미지 업로드</h1>
-    <form action="/receipt_OCR" method="POST" enctype="multipart/form-data">
-      <input type="file" name="image" />
-      <input type="submit" value="업로드" />
-    </form>
-  </body>
-</html>
